+++ include/oc_blockwise.h
@@ -76,19 +76,19 @@ oc_blockwise_state_t *oc_blockwise_find_response_buffer_by_client_cb(
   oc_endpoint_t *endpoint, void *client_cb);
 
 oc_blockwise_state_t *oc_blockwise_find_request_buffer(
-  const char *href, int href_len, oc_endpoint_t *endpoint, oc_method_t method,
+  const char *href, size_t href_len, oc_endpoint_t *endpoint, oc_method_t method,
   const char *query, int query_len, oc_blockwise_role_t role);
 
 oc_blockwise_state_t *oc_blockwise_find_response_buffer(
-  const char *href, int href_len, oc_endpoint_t *endpoint, oc_method_t method,
-  const char *query, int query_len, oc_blockwise_role_t role);
+  const char *href, size_t href_len, oc_endpoint_t *endpoint, oc_method_t method,
+  const char *query, size_t query_len, oc_blockwise_role_t role);
 
 oc_blockwise_state_t *oc_blockwise_alloc_request_buffer(
-  const char *href, int href_len, oc_endpoint_t *endpoint, oc_method_t method,
+  const char *href, size_t href_len, oc_endpoint_t *endpoint, oc_method_t method,
   oc_blockwise_role_t role);
 
 oc_blockwise_state_t *oc_blockwise_alloc_response_buffer(
-  const char *href, int href_len, oc_endpoint_t *endpoint, oc_method_t method,
+  const char *href, size_t href_len, oc_endpoint_t *endpoint, oc_method_t method,
   oc_blockwise_role_t role);
 
 void oc_blockwise_free_request_buffer(oc_blockwise_state_t *buffer);
